+++ src/hooks/useCenteringModes/index.ts
@@ -1,43 +1,1 @@
-import { useCallback, useMemo } from 'react';
-import { useDispatch } from 'react-redux';
-import { useTranslation } from 'react-i18next';
-
-import { useTypedSelector } from 'hooks/useTypedSelector';
-
-import * as Actions from 'store/actions';
-
-import { CenteringMode } from 'enums/CenteringMode';
-
-type SetCenteringModeCallback = (centeringMode: CenteringMode) => void;
-
-export const useCenteringModes = () => {
-  const { t } = useTranslation();
-
-  const dispatch = useDispatch();
-
-  const { activeCenteringMode } = useTypedSelector(
-    (state) => state.centeringModes
-  );
-
-  const centeringModes = useMemo(
-    () => [
-      { name: t('settings.centeringMode.center'), value: CenteringMode.Center },
-      { name: t('settings.centeringMode.fill'), value: CenteringMode.Fill },
-      { name: t('settings.centeringMode.none'), value: CenteringMode.None },
-    ],
-    [t]
-  );
-
-  const setCenteringMode = useCallback<SetCenteringModeCallback>(
-    (centeringMode) => {
-      dispatch(Actions.setActiveCenteringMode(centeringMode));
-    },
-    [dispatch]
-  );
-
-  return {
-    centeringModes,
-    activeCenteringMode,
-    setCenteringMode,
-  };
-};
+export { useCenteringModes } from './useCenteringModes';
