+++ src/hooks/useSupportedScreenSize/index.ts
@@ -1,14 +1,1 @@
-import { useMemo } from 'react';
-import { useViewport } from 'react-viewport-hooks';
-
-const MIN_WINDOW_SIZE = 960;
-
-export const useSupportedScreenSize = () => {
-  const { vw } = useViewport({
-    defaultVW: window.innerWidth,
-  });
-
-  const isSupportedScreenSize = useMemo(() => vw > MIN_WINDOW_SIZE, [vw]);
-
-  return isSupportedScreenSize;
-};
+export { useSupportedScreenSize } from './useSupportedScreenSize';
