+++ src/hooks/useSupportedScreenSize/useSupportedScreenSize.ts
@@ -1,0 +1,14 @@
+import { useMemo } from 'react';
+import { useViewport } from 'react-viewport-hooks';
+
+const MIN_WINDOW_SIZE = 960;
+
+export const useSupportedScreenSize = () => {
+  const { vw } = useViewport({
+    defaultVW: window.innerWidth,
+  });
+
+  const isSupportedScreenSize = useMemo(() => vw > MIN_WINDOW_SIZE, [vw]);
+
+  return isSupportedScreenSize;
+};
