+++ src/hooks/useTypedSelector/index.ts
@@ -1,5 +1,1 @@
-import { TypedUseSelectorHook, useSelector } from 'react-redux';
-
-import { RootState } from 'store/reducers';
-
-export const useTypedSelector: TypedUseSelectorHook<RootState> = useSelector;
+export { useTypedSelector } from './useTypedSelector';
