+++ src/hooks/useTypedSelector/useTypedSelector.ts
@@ -1,0 +1,5 @@
+import { TypedUseSelectorHook, useSelector } from 'react-redux';
+
+import { RootState } from 'store/reducers';
+
+export const useTypedSelector: TypedUseSelectorHook<RootState> = useSelector;
